requirements.txt
boto3==1.17.17
botocore==1.20.17
python-dotenv==0.15.0